refactoring code to remove duplication
